move roomQuantitySelect outside of Navigation
